Code refactoring for peers
